Created a separate component for ParameterBox
